main.c
Initial commit of the code. I've been playinga around with it quite a bit so it's a mess here. It also relies on code in an avr library on my personal computer that needs to be commited. So with this initial commit. The code wont' work. This is really just to back up what I do have on my machine.